Clean up initialisation
