--- src/app/calendar/calendar.tpl.html
@@ -1,88 +1,0 @@
-<div class="wfm-mobile-calendar" flex>
-
-<md-datepicker ng-model="myDate" md-placeholder="Enter date"></md-datepicker>
-
-<table class="wfm-mobile-calendar-table">
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td>
-      <div class="wfm-mobile-calendar-entry">
-        Example of text
-      </div>
-    </td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td>
-    </td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td>
-      <div class="wfm-mobile-calendar-entry" style="height:70px;">
-        Example of text
-      </div>
-    </td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-  <tr>
-    <td class="wfm-mobile-calendar-time">
-      10 am
-    </td>
-    <td></td>
-  </tr>
-</table>
-</div>
