+++ src/app/initialisation/index.js
@@ -1,0 +1,3 @@
+require('./syncPoolService');
+require('./config');
+require('./run');
