+++ src/app/main.tpl.html
@@ -30,10 +30,6 @@
         <md-icon md-font-set="material-icons">map</md-icon>
         <p>Map</p>
       </md-list-item>
-      <md-list-item ng-click="navigateTo('app.calendar'); toggleSidenav($event, 'left')" ng-class="{active: $state.includes('app.calendar')}">
-        <md-icon md-font-set="material-icons">today</md-icon>
-        <p>Calendar</p>
-      </md-list-item>
       <md-list-item ng-click="navigateTo('app.file'); toggleSidenav($event, 'left')" ng-class="{active: $state.includes('app.file')}">
         <md-icon md-font-set="material-icons">folder</md-icon>
         <p>Files</p>
